# 8321620: Optimize JImage decompressors