--- a/src/java.base/share/classes/jdk/internal/jimage/decompressor/StringSharingDecompressor.java
+++ b/src/java.base/share/classes/jdk/internal/jimage/decompressor/StringSharingDecompressor.java
@@ -177,8 +177,8 @@ public class StringSharingDecompressor implements ResourceDecompressor {
                 argIndex += 1;
                 String pkg = reader.getString(index);
                 if (!pkg.isEmpty()) {
-                    pkg = pkg + "/";
                     current += ImageStringsReader.mutf8FromString(bytesOut, current, pkg);
+                    bytesOut[current++] = '/';
                 }
                 int classIndex = indices[argIndex];
                 argIndex += 1;