… test/jdk/tools/jlink/plugins/StringSharingPluginTest.java