@Glavo 2023-11-08
I generated the runtime image using `jlink --compress (1|2) --add-modules java.se,jdk.unsupported,jdk.management` and then ran the following JMH benchmark:

```java
@Warmup(iterations = 5, time = 2)
@Measurement(iterations = 5, time = 2)
@Fork(value = 1, jvmArgsAppend = {"-XX:+UseG1GC", "-Xms8g", "-Xmx8g", "--add-exports=java.base/jdk.internal.jimage=ALL-UNNAMED"})
@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@State(Scope.Benchmark)
public class Decompress {

    private static final ImageReader READER = ImageReaderFactory.getImageReader();
    private static final ImageLocation LOC = READER.findLocation("java.base", "java/lang/String.class");

    @Benchmark
    public ByteBuffer test() {
        return READER.getResourceBuffer(LOC);
    }

}
```

This is the result:

```
String Share
                                                            (Baseline)                        (Current)
Benchmark                           Mode  Cnt       Score       Error   Units         Score      Error   Units
Decompress.test                     avgt    5  256828.072 ± 10219.803   ns/op     74216.942 ± 1858.956   ns/op   (-71.10%)
Decompress.test:gc.alloc.rate       avgt    5    2341.441 ±    92.021  MB/sec      2826.574 ±   70.496  MB/sec
Decompress.test:gc.alloc.rate.norm  avgt    5  630544.882 ±     0.035    B/op    219976.255 ±    0.006    B/op   (-65.11%)
Decompress.test:gc.count            avgt    5       5.000              counts         5.000             counts
Decompress.test:gc.time             avgt    5      11.000                  ms        11.000                 ms


Zip

Benchmark                           Mode  Cnt       Score       Error   Units        Score       Error   Units
Decompress.test                     avgt    5  194237.534 ±  1026.180   ns/op   152855.728 ± 16058.780   ns/op   (-21.30%)
Decompress.test:gc.alloc.rate       avgt    5    1197.700 ±     6.306  MB/sec      464.278 ±    47.465  MB/sec
Decompress.test:gc.alloc.rate.norm  avgt    5  243953.338 ±     5.810    B/op    74376.291 ±     2.175    B/op   (-69.51%)
Decompress.test:gc.count            avgt    5       2.000              counts        1.000              counts
Decompress.test:gc.time             avgt    5       4.000                  ms        3.000                  ms
```

The results show that memory allocation is reduced by more than 60% while decompression speed is significantly improved.