@Glavo 2023-11-09
I generated a runtime image containing javac using `jlink --compress (0|1) --add-modules jdk.compiler` and tested the time it took to compile a simple HelloWorld program 20 times using the `perf stat -r20 javac /dev/shm/HelloWorld.java` command. This is the result:

Baseline, No Compress: 10829ms

String Share:

* Baseline: 17830ms
* This PR: 13598ms

Zip:

* Baseline: 12350ms
* This PR: 12279ms

You can see that in this test, this PR made the runtime image compressed using string share 30% faster.