@cl4es 2023-11-20
I'll need some time to review this thoroughly but skimming through it it seems like a very nice enhancement, making `--compress` more appealing - great work! 

One gotcha in this area is that `jlink` by default does not recreate the CDS archive. Startup might gain from generating out the default CDS archive (`java -Xshare:dump`), but not doing so is a reasonable thing to do if you're aiming for minimal static footprint. To make comparisons fair you need to take care to compare apples to apples.

(I'll need some more time for a formal review)