@Glavo 2023-11-23
> But to make comparisons fair you need to take care to compare apples to apples: is the uncompressed baseline a similar jlink image generated without --compress? Or the default JDK image (uncompressed, with CDS)?

The uncompressed runtime image used for comparison was also generated using jlink.