@cl4es 2023-12-08
> Does anyone want to take a look at this PR?

RFE: https://bugs.openjdk.org/browse/JDK-8321620 - update the bug ID and this should PR should reach a wider audience. I'll have some time next week to help review.