@Glavo 2024-01-12
@cl4es 

> While the improvement to --compress 1 are impressive, this compression mode is effectively deprecated.

I noticed the deprecation warning before creating this PR, so I did some research.

This warning was introduced here: https://github.com/openjdk/jdk/pull/11617. In the discussion of that PR I learned that the option `--compress 1` was deprecated because it was not compression but string sharing. The string sharing plugin is not actually deprecated, it's just that it may be renamed in the future.