@Glavo 2024-01-12
> The plan [1] is to remove the old compression values `0`, `1`, `2` in the future and only support the zip compression `--compress zip-[0-9]`, i.e. the string sharing plugin will be removed as there isn't any known customer usage of this plugin. If you are aware of any customer using it, we can reevaluate.
> 
> [1] https://bugs.openjdk.org/browse/JDK-8301124?focusedId=14562770&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-14562770

I expected that few people use this option, because compared to zip, it is not superior in file size or decompression speed. However, I did some experiments in the [JApp](https://github.com/Glavo/japp) project and found that the file size using a mixture of string sharing + zstd compression was significantly smaller than using only zstd, so I thought it might be meaningful to study it further.

Anyway, these are things for the future and I want this PR to stay simple, so I'll just drop these modifications.