@Glavo 2024-01-12
I created a new PR: #17405

I want to leave my changes to the string sharing plugin here so I can pick them up in the future.